Updated relative and absolute positioning & utilities
The way relativity is handled was updated in a way that is breaking for complex inventories.

Every default implementation of InventoryContentView is now relative to their parents. Such that their section is relative to their parent.

For more on this read the documentation for InventorySection#toRelative(InventoryPosition) and InventorySection#toAbsolute